+++ chia/apis/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from chia.apis.api_protocol_registry import ApiProtocolRegistry
+from chia.apis.harvester_stub import HarvesterApiStub
+
+__all__ = ["ApiProtocolRegistry", "HarvesterApiStub"]
